pyproject.toml
[project]
name = "mcp-gemini-websearch"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.14"
dependencies = [
    "asyncpg>=0.31.0",
    "fastmcp>=3.1.1",
    "google-genai>=1.68.0",
]

[dependency-groups]
dev = [
    "pytest-asyncio>=1.3.0",
]
